takimlar eklendi ama statik,onlari otomatik yap.yeni team eklenirse patlar.tabloya scroll x eklendi. aciklama eklendi. double remaning time int yapildi ve KRITIK BUG DUZELTILDI. Takimlarda capacity set edili degilse GSON patliyordu

+++ src/main/java/model/pojo/TempoTeams/Team.java
@@ -26,13 +26,13 @@ public class Team {
     private Boolean isPublic;
     @SerializedName("RemaningInAYear")
     @Expose
-    private Double RemainingInAYear;
+    private Integer RemainingInAYear;
 
-    public void setRemainingInAYear(Double RemainingInAYear) {
+    public void setRemainingInAYear(Integer RemainingInAYear) {
         this.RemainingInAYear = RemainingInAYear;
     }
 
-    public Double getRemainingInAYear() {
+    public Integer getRemainingInAYear() {
         return RemainingInAYear;
     }
